requirements.txt
pandas
dotenv
datasets
transformers
filelock
openai
matplotlib
numpy
google-genai
